refactor(spf)!: split background-video into engine and adapter
The Medias moved to `playback/adapters/` and gained their own entry points so that
driving an engine directly costs neither a Media nor `@videojs/media`.
`background-video` predated that split: it kept both halves in
`playback/engines/` and shipped them from one `@videojs/spf/background-video`
entry, leaving the only adapter still sitting in the engine layer.

Put it on the same footing. The adapter moves to `playback/adapters/` and keeps
`@videojs/spf/background-video`; the engine moves beside the other HLS engines as
`engine-background-video.ts` and ships from `@videojs/spf/hls`. That second half
follows `engine-audio-only`, which is also a subtractive variant of the full HLS
engine and has always lived there — engines share one entry per format family,
adapters get one each. So no new entry point, and nothing outside the package
changes but the sandbox's engine-type import.

No host-bound Media is introduced. Nothing consumes one, and binding the mixin to
`HTMLVideoElementHost` would give this entry a `@videojs/media` dependency it does
not otherwise have; `BackgroundVideoMediaMixin` composes over any base when that
changes.

Folding the third engine into `./hls` costs 281 B gzipped, taking it to 19.36 KB
against its 20 KB target.

The use-case record also named a picker symbol, `pickMaxResolutionVideoTrack`,
that no longer exists in the source; corrected to the shipped primitives.
