--- packages/spf/src/playback/engines/background-video/index.ts
@@ -1,13 +1,0 @@
-export type { BackgroundVideoMediaAPI, BackgroundVideoMediaProps } from './adapter';
-export {
-  BackgroundVideoMediaElement,
-  BackgroundVideoMediaMixin,
-  backgroundVideoMediaDefaultProps,
-} from './adapter';
-export type {
-  BackgroundVideoEngineConfig,
-  BackgroundVideoEngineContext,
-  BackgroundVideoEngineSignals,
-  BackgroundVideoEngineState,
-} from './engine';
-export { createBackgroundVideoEngine } from './engine';
